Updated store-report command to a multimethod
We will need to support multiple versions of the store-report
command for Burgundy since we're changing the wire format.

This commit simply re-works the existing store report command
to use a multimethod, and introduces v1 of the multimethod.
This commit does not introduce a v2 implementation of the
multimethod yet.