Add support for `count-by` query parameter
Prior to this commit, all of the count operations were based on the
number of *resources* that matched the event queries.  However,
some of the Burgundy requirements necessitated that we be able
to count based on the number of *nodes* that matched the event
query.  This commit adds a `count-by` query parameter that can be set
to either `resource` or `node`.  If it is set to `node`, then the
event query results are filtered through a DISTINCT clause that
limits the result set to unique occurrences per node.