(PE-962) Add schema for flagging latest report per node
This commit adds a new table for storing data about which
report was the most recent for each node.  This will make
it easier to write event queries that only return events
from the last puppet run for each node.

The migration code also has to populate this new table for
all of the existing data.  The commit includes tests
to validate that migration code.

Conflicts:

	src/com/puppetlabs/puppetdb/scf/migrate.clj