Fabricate failure events for failed resource_statuses
Prior to this commit, it was possible for puppet to generate a
report where there was a resource_status with `failed` == `true`,
but no events in the events array.  The PuppetDB terminus
assumed that there would always be an event representing the
failure, so it did not generate any events for this.  The
result is that things that were clear failures in puppet
might not show up at all in the event inspector.

This commit changes the report processor so that we treat
`failed` resource_statuses who have no events in their array
the same way that we do for `skipped` resource_statuses.
This will cause them to show up as events in puppetdb.